pin dialog component
